move time in lib
